requirements.txt
pvporcupine
pyaudio
python-dotenv
torch
torchaudio
transformers
accelerate
soundfile
pyttsx3
vosk